Restructuration du code pour plus de lisibilité
